@@ -1,69 +1,150 @@
-using AppG.BBDD.Respuestas;
-using AppG.Entidades.BBDD;
-using AppG.Servicio;
+﻿using AhorroLand.Application.Features.Personas.Commands;
+using AhorroLand.Application.Features.Personas.Queries;
+using AhorroLand.Application.Features.Personas.Queries.Recent;
+using AhorroLand.NuevaApi.Controllers.Base;
+using AhorroLand.Shared.Domain.Abstractions.Results; // Para Error y Result
+using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.OutputCaching;
 
-namespace AppG.Controllers
+namespace AhorroLand.NuevaApi.Controllers;
+
+[Authorize]
+[ApiController]
+[Route("api/personas")]
+public class PersonasController : AbsController
 {
-    [ApiController]
-    [Route("api/persona")]
-    [Authorize]
-    public class PersonaController : BaseController<Persona>
+    public PersonasController(ISender sender) : base(sender)
     {
-        private readonly IPersonaServicio _personaService;
+    }
 
-        public PersonaController(IPersonaServicio personaService) : base(personaService)
+    /// <summary>
+    /// Obtiene lista paginada de personas del usuario autenticado.
+    /// </summary>
+    [HttpGet]
+    public async Task<IActionResult> GetAll([FromQuery] int page = 1, [FromQuery] int pageSize = 10)
+    {
+        // 1. Obtener ID del usuario (Seguridad)
+        var usuarioId = GetCurrentUserId();
+
+        if (usuarioId is null)
         {
-            _personaService = personaService;
+            return Unauthorized(Result.Failure(Error.Unauthorized("Usuario no autenticado")));
         }
 
-        [HttpGet("getPersonas/{idUsuario}")]
-        public async Task<IActionResult> GetPersonas(int idUsuario)
+        // 2. Crear query filtrando por usuario
+        var query = new GetPersonasPagedListQuery(page, pageSize)
         {
-            var result = await _personaService.GetAllAsync(idUsuario);
+            UsuarioId = usuarioId.Value // 👈 IMPORTANTE: Asignar el ID
+        };
 
-            if (result is IDictionary<string, object> errorResult && errorResult.ContainsKey("Error"))
-            {
-                return StatusCode(500, errorResult);
-            }
+        var result = await _sender.Send(query);
+        return HandleResult(result);
+    }
 
-            return Ok(result);
+    /// <summary>
+    /// Búsqueda rápida para autocomplete.
+    /// </summary>
+    [HttpGet("search")]
+    public async Task<IActionResult> Search([FromQuery] string search, [FromQuery] int limit = 10)
+    {
+        var usuarioId = GetCurrentUserId();
+
+        if (usuarioId is null)
+        {
+            return Unauthorized(Result.Failure(Error.Unauthorized("Usuario no autenticado")));
         }
 
-
-        [HttpGet("getCantidad")]
-        public async Task<IActionResult> GetCantidad(int page, int size, int idUsuario)
+        var query = new SearchPersonasQuery(search, limit)
         {
-            var result = await _personaService.GetCantidadAsync(page, size, idUsuario);
+            UsuarioId = usuarioId.Value
+        };
 
-            if (result is IDictionary<string, object> errorResult && errorResult.ContainsKey("Error"))
-            {
-                return StatusCode(500, errorResult);
-            }
+        var result = await _sender.Send(query);
+        return HandleResult(result);
+    }
 
-            return Ok(result);
+    /// <summary>
+    /// Obtiene las personas más recientes.
+    /// </summary>
+    [HttpGet("recent")]
+    public async Task<IActionResult> GetRecent([FromQuery] int limit = 5)
+    {
+        var usuarioId = GetCurrentUserId();
+
+        if (usuarioId is null)
+        {
+            return Unauthorized(Result.Failure(Error.Unauthorized("Usuario no autenticado")));
         }
 
-        public override async Task<IActionResult> Create([FromBody] Persona entity)
+        var query = new GetRecentPersonasQuery(limit)
         {
+            UsuarioId = usuarioId.Value
+        };
 
-            var createdEntity = await _personaService.CreateAsync(entity);
+        var result = await _sender.Send(query);
+        return HandleResult(result);
+    }
 
-            var message = $"{typeof(Persona).Name} creado correctamente";
+    [HttpGet("{id}")]
+    public async Task<IActionResult> GetById(Guid id)
+    {
+        var query = new GetPersonaByIdQuery(id);
+        var result = await _sender.Send(query);
+        return HandleResult(result);
+    }
 
-            var response = new ResponseOne<Persona>(createdEntity, message);
+    [HttpPost]
+    public async Task<IActionResult> Create([FromBody] CreatePersonaRequest request)
+    {
+        // Asignación inteligente de UsuarioId
+        var usuarioId = request.UsuarioId != Guid.Empty ? request.UsuarioId : GetCurrentUserId() ?? Guid.Empty;
 
-            return Ok(response);
-        }
-
-        public override async Task<IActionResult> Update(int id, [FromBody] Persona entity)
+        var command = new CreatePersonaCommand
         {
+            Nombre = request.Nombre,
+            UsuarioId = usuarioId
+        };
 
-            await _personaService.UpdateAsync(id, entity);
+        var result = await _sender.Send(command);
 
-            return Ok(new { message = $"{typeof(Persona).Name} actualizado correctamente" });
+        // Uso seguro de HandleResultForCreation
+        return HandleResultForCreation(
+            result,
+            nameof(GetById),
+            new { id = result.IsSuccess ? result.Value : Guid.Empty }
+        );
+    }
 
-        }
+    [HttpPut("{id}")]
+    public async Task<IActionResult> Update(Guid id, [FromBody] UpdatePersonaRequest request)
+    {
+        var command = new UpdatePersonaCommand
+        {
+            Id = id,
+            Nombre = request.Nombre
+        };
+
+        var result = await _sender.Send(command);
+        return HandleResult(result);
+    }
+
+    [HttpDelete("{id}")]
+    public async Task<IActionResult> Delete(Guid id)
+    {
+        var command = new DeletePersonaCommand(id);
+        var result = await _sender.Send(command);
+        return HandleResult(result);
     }
 }
+
+// DTOs
+public record CreatePersonaRequest(
+    string Nombre,
+    Guid UsuarioId
+);
+
+public record UpdatePersonaRequest(
+    string Nombre
+);
