@@ -1,68 +1,151 @@
-using AppG.BBDD.Respuestas;
-using AppG.Entidades.BBDD;
-using AppG.Servicio;
+﻿using AhorroLand.Application.Features.Proveedores.Commands;
+using AhorroLand.Application.Features.Proveedores.Queries;
+using AhorroLand.Application.Features.Proveedores.Queries.Recent;
+using AhorroLand.Application.Features.Proveedores.Queries.Search; // Asegúrate de tener este namespace
+using AhorroLand.NuevaApi.Controllers.Base;
+using AhorroLand.Shared.Domain.Abstractions.Results; // Para Error y Result
+using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.OutputCaching;
 
-namespace AppG.Controllers
+namespace AhorroLand.NuevaApi.Controllers;
+
+[Authorize]
+[ApiController]
+[Route("api/proveedores")]
+public class ProveedoresController : AbsController
 {
-    [ApiController]
-    [Route("api/proveedor")]
-    [Authorize]
-    public class ProveedorController : BaseController<Proveedor>
+    public ProveedoresController(ISender sender) : base(sender)
     {
-        private readonly IProveedorServicio _proveedorService;
+    }
 
-        public ProveedorController(IProveedorServicio proveedorService) : base(proveedorService)
+    /// <summary>
+    /// Obtiene lista paginada de proveedores del usuario autenticado.
+    /// </summary>
+    [HttpGet]
+    public async Task<IActionResult> GetAll([FromQuery] int page = 1, [FromQuery] int pageSize = 10)
+    {
+        // 1. Obtener ID del usuario (Seguridad)
+        var usuarioId = GetCurrentUserId();
+
+        if (usuarioId is null)
         {
-            _proveedorService = proveedorService;
+            return Unauthorized(Result.Failure(Error.Unauthorized("Usuario no autenticado")));
         }
 
-        [HttpGet("getProveedores/{idUsuario}")]
-        public async Task<IActionResult> GetProveedores(int idUsuario)
+        // 2. Crear query filtrando por usuario
+        var query = new GetProveedoresPagedListQuery(page, pageSize)
         {
-            var result = await _proveedorService.GetAllAsync(idUsuario);
+            UsuarioId = usuarioId.Value // 👈 IMPORTANTE: Asignar el ID
+        };
 
-            if (result is IDictionary<string, object> errorResult && errorResult.ContainsKey("Error"))
-            {
-                return StatusCode(500, errorResult);
-            }
+        var result = await _sender.Send(query);
+        return HandleResult(result);
+    }
 
-            return Ok(result);
+    /// <summary>
+    /// Búsqueda rápida para autocomplete.
+    /// </summary>
+    [HttpGet("search")]
+    public async Task<IActionResult> Search([FromQuery] string search, [FromQuery] int limit = 10)
+    {
+        var usuarioId = GetCurrentUserId();
+
+        if (usuarioId is null)
+        {
+            return Unauthorized(Result.Failure(Error.Unauthorized("Usuario no autenticado")));
         }
 
-        [HttpGet("getCantidad")]
-        public async Task<IActionResult> GetCantidad(int page, int size, int idUsuario)
+        var query = new SearchProveedoresQuery(search, limit)
         {
-            var result = await _proveedorService.GetCantidadAsync(page, size, idUsuario);
+            UsuarioId = usuarioId.Value
+        };
 
-            if (result is IDictionary<string, object> errorResult && errorResult.ContainsKey("Error"))
-            {
-                return StatusCode(500, errorResult);
-            }
+        var result = await _sender.Send(query);
+        return HandleResult(result);
+    }
 
-            return Ok(result);
+    /// <summary>
+    /// Obtiene los proveedores más recientes.
+    /// </summary>
+    [HttpGet("recent")]
+    public async Task<IActionResult> GetRecent([FromQuery] int limit = 5)
+    {
+        var usuarioId = GetCurrentUserId();
+
+        if (usuarioId is null)
+        {
+            return Unauthorized(Result.Failure(Error.Unauthorized("Usuario no autenticado")));
         }
 
-        public override async Task<IActionResult> Create([FromBody] Proveedor entity)
+        var query = new GetRecentProveedoresQuery(limit)
         {
+            UsuarioId = usuarioId.Value
+        };
 
-            var createdEntity = await _proveedorService.CreateAsync(entity);
+        var result = await _sender.Send(query);
+        return HandleResult(result);
+    }
 
-            var message = $"{typeof(Proveedor).Name} creado correctamente";
+    [HttpGet("{id}")]
+    public async Task<IActionResult> GetById(Guid id)
+    {
+        var query = new GetProveedorByIdQuery(id);
+        var result = await _sender.Send(query);
+        return HandleResult(result);
+    }
 
-            var response = new ResponseOne<Proveedor>(createdEntity, message);
+    [HttpPost]
+    public async Task<IActionResult> Create([FromBody] CreateProveedorRequest request)
+    {
+        // Asignación inteligente de UsuarioId
+        var usuarioId = request.UsuarioId != Guid.Empty ? request.UsuarioId : GetCurrentUserId() ?? Guid.Empty;
 
-            return Ok(response);
-        }
-
-        public override async Task<IActionResult> Update(int id, [FromBody] Proveedor entity)
+        var command = new CreateProveedorCommand
         {
+            Nombre = request.Nombre,
+            UsuarioId = usuarioId
+        };
 
-            await _proveedorService.UpdateAsync(id, entity);
+        var result = await _sender.Send(command);
 
-            return Ok(new { message = $"{typeof(Proveedor).Name} actualizado correctamente" });
-        }
+        // Uso seguro de HandleResultForCreation
+        return HandleResultForCreation(
+            result,
+            nameof(GetById),
+            new { id = result.IsSuccess ? result.Value : Guid.Empty }
+        );
+    }
 
+    [HttpPut("{id}")]
+    public async Task<IActionResult> Update(Guid id, [FromBody] UpdateProveedorRequest request)
+    {
+        var command = new UpdateProveedorCommand
+        {
+            Id = id,
+            Nombre = request.Nombre
+        };
+
+        var result = await _sender.Send(command);
+        return HandleResult(result);
+    }
+
+    [HttpDelete("{id}")]
+    public async Task<IActionResult> Delete(Guid id)
+    {
+        var command = new DeleteProveedorCommand(id);
+        var result = await _sender.Send(command);
+        return HandleResult(result);
     }
 }
+
+// DTOs
+public record CreateProveedorRequest(
+    string Nombre,
+    Guid UsuarioId
+);
+
+public record UpdateProveedorRequest(
+    string Nombre
+);
